feat(network): proxy third-party API calls in dev env
It allows a third-party API (e.g. served on http://localhost:8000) to use cookie authentication even if the Next.js app is served on another domain (e.g. http://localhost:3000). Without proxy http://localhost:8000 is seen as third-party by the browser and won't be able to use cookies except if the browser allows third-party cookies (which is only the case by default on Chrome, and even then it's supposed to change "soon").
You should not need that proxy in production env if the API is on a subdomain of the Next.js app, as long as the cookie `Domain` flag is set to the domain of the Next.js app.